+++ src/ios/OSInAppBrowserError.swift
@@ -1,6 +1,7 @@
 enum OSInAppBrowserTarget {
-    case openInExternalBrowser
-    case openInSystemBrowser
+    case externalBrowser
+    case systemBrowser
+    case webView
 }
 
 enum OSInAppBrowserError: Error {
@@ -26,8 +27,9 @@ enum OSInAppBrowserError: Error {
             let targetString: String
             
             switch target {
-            case .openInExternalBrowser: targetString = "openInExternalBrowser"
-            case .openInSystemBrowser: targetString = "openInSystemBrowser"
+            case .externalBrowser: targetString = "openInExternalBrowser"
+            case .systemBrowser: targetString = "openInSystemBrowser"
+            case .webView: targetString = "openInWebView"
             }
             
             result = "The input parameters for '\(targetString)' are invalid."
@@ -35,8 +37,9 @@ enum OSInAppBrowserError: Error {
             let targetString: String
             
             switch target {
-            case .openInExternalBrowser: targetString = "Safari"
-            case .openInSystemBrowser: targetString = "SFSafariViewController"
+            case .externalBrowser: targetString = "Safari"
+            case .systemBrowser: targetString = "SFSafariViewController"
+            case .webView: targetString = "WebView"
             }
             
             result = "Couldn't open '\(url)' using \(targetString)."